requirements.txt
pyyaml==5.4.1
opencv-python==4.5.3.56
opencv-python==4.5.3.56
pathlib==1.0.1
optuna==2.9.1
pandas==1.3.2
numpy==1.18.5
natsort==7.1.1
mlflow==1.20.2
ipykernel==6.2.0
dvc==2.7.4
tensorflow==2.3.0
